Kinetic: Do not `git submodule' Kinetic/kinetic-protocol
The Kinetic code depends on the `kinetic.proto` declaration file, which
is sourced from a git submodule. Problem: when Arsenal / Kinetic is
installed with npm (as a dependency for another project such as
IronMan-Data), submobules are not cloned and `kinetic.proto` is missing.

Since we only need the `kinetic.proto` file from
Kinetic/kinetic-protocol, and this specification is not supposed to
change, let's copy it in our repo and drop the annoying submodule
dependency.