Kinetic: Simulator: Cache maven and .jar files for CI
Downloading and building the simulator is a long step (~ minutes)
because there are lots of .jar files to download and compose with.
This causes builds to last much longer than before. To palliate this,
let's add a `build_cache` Makefile directive and call it in an early
stage of the build, so CircleCI can cache it for future builds.

This enhancement reduce `npm run tests_functional_simulator` duration
from 174s to 3s.