Kinetic: Define PDU commands as classes
Currently to create a "PUT" Kinetic PDU, one must run:

    const pdu = new Kinetic.PDU();
    pdu.put();

This is not object-oriented programming, and is error-prone since
we don't expect commands like `put`, `flush` etc. to create messages
with a given type.

This patch creates classes that inherit from `PDU`, e.g. `PutPDU`,
`GetPDU`, `GetLogPDU` etc. and that can be used the same way.