bf: convert sync calls to async
The in-memory backend emulators are being used assuming they are async
when they are actually synchronous. Wrapping these calls with async.js
results in 'Maximum call stack size exceeded' error as all functions
are called on the same tick. Wrapping them with process.nextTick()
ensures they are called on the next tick of the event loop.

Fix #527