COMPAT: Return expected response for `MaxUpload` and `Delimiter` params
Currently, the value of `MaxUploads` for listMultipartUpload API is
the number of MPU in a bucket. This is not the behavior of AWS. This
commit changes the API's behavior so the value of `MaxUploads` can be
specified by the user using the `MaxUploads` parameter. This commit also
returns the expected default of 1000 if no `MaxUploads` is not defined.
If `MaxUploads` is defined as `0`, the behavior of AWS is to return an
empty `Uploads` array and set `IsTruncated` to `false`. This commit also
adds this expected behavior.

Currently, when using the file-backend, if a `Delimiter` parameter is
defined and it does not result in a `CommonPrefixes` array that is
greater than 0, `Delimiter` is set to ''. This causes the XML response
to omit the `Delimeter` key. For compatibility with AWS, this commit
tests an associated change in Arsenal that fixes the `Delimiter` response.

A new test has been added to verify these behaviors.

[squash me] Use object destructuring for variables

[squash me] Clean up code and comments

[squash me] Remove callbacks and return all every Promise

[squash me] Handle 0 `MaxUploads` object in `getExpectedObj()`

[squash me] Separate assertions from object creation

[squash me] Remove redundant assertion statements