listing: handle non alphabetical order in delimiter.
For the S3SOFS feature, the listing results are not albetically sorted
so we can't rely on a basic string comparison to know if we should skip
a value when filtering a listing result.

This commit adds a albeticalOrder field to the Delimiter class. It is
set to true by default, avoiding to break anything.