ft: update models
Because I did not know the merge process in the Arsenal repository I
merged too early a PR moving code from S3 to Arsenal. The problem is
that the PR in S3 removing this models is not yet merged and some
modification have been mode on this code in S3, leading the previous S3
PR to fail.

This commit updates the models code and test in Arsenal, accordingly to
what has been done in S3.
In the S3 repository, the ReplicationConfiguration source file required
the S3 config. Because it is not possible to do that in the Arsenal
repository, the config is now a constructor parameter.