[Tests][Algos/list] Delimiter: Ensure that listing stops beyond prefix
Some bugs proved that the fact that we're not explicitly stopping the listing
once we reach the boundary of the provided prefix parameter may lead to
inconsistent behaviors.

This commit adds a test that shows that the current code is not stopping once
the keys are greater than the provided prefix parameter, thus skipping entries
that need not even be read with a correct approach.

Relates to MD-654