ZENKO-1124: fix delimiterMaster::filter delete markers handlings.
A specific dataset described and attached in the linked Jira issue shows
a bug in that the current delimiterMaster::filter handling of the delete marker
on a master version and the use of this function into MongoDB client
leads to an infinite loop.

The input dataset, in a versionedis in the form:
- one object with a master version and some others,
- 10k objects with a delete marker on the master version and some other
  versions for each object.
The expected result for this input contains only the first object.

What happens in the delimiterMaster:
- filtering the first object results to adds it to the result and set
  NextMarker to its key. The following versions of this objets are
  skipped.
- filtering the others objets results to skip return values.

In the MongoDB client class, after counting 100 skipped values in a row,
the skipscan mechanism is triggered. The problem is that the new range
is computed using NextMarker, which in this case is set to the key of
the 1st object. The listing is done on the same range (well, the first
object is skipped at least).

The fix is to set the NextMarker when skipping a delete marker on a
master version. It will allow the client code to effectively skip the
range. As far as I see there is no problem to set this field in a lot of
other cases (it is only set when the entry is not skipped) but it may be
risky to modify the internal of this function and it is not the topic of
this urgent fix.