S3C-1665: delimiter*. return FILTER_END ASAP
For all the delimiter extension, once we reach the maxKey parameter
entries added to the result, we should return FILTER_END on the next
entry.

Currently, we wait to find the next entry to add to the result to
return FILTER_END (and we don't add this entry to the result).

This commit moves the check to test is maxKeys is reached from
addContent/addCommon methods to the begining of the filter method.