feature: ZENKO-1520 expose prometheus metrics in HealthProbeServer
Add a new route '/_/monitoring/metrics' exposed by the
HealthProbeServer that runs on pods, to expose default metrics (nodejs
etc) and custom metrics in prometheus format. Then, prometheus will be
able to scrape them for each pod.

Ideally the class should be renamed, maybe to MonitoringServer, kept
it for later as it involves a larger refactor.