bf: ZENKO-1768 delimiterMaster similar key listing
Given the following situation:
- master isPHD and key name for example is "test1"
- version "test1"
- master isPHD and key name for example is "test10"
- version "test10"

We were filtering out "test10" from listings even though it
was considered a valid entry in the listing.

There was a check made where given the above scenario, we
would check if `this.delimiter` exists, then see if current
key startsWith the previous continue marker. Instead, we
should check if the current key startsWith the previous
continue marker + `this.delimiter`.