improvement: ZENKO-2535 add microVersionId to ObjectMD
Add a new microVersionId field that is a hex-encoded field of 64 bits
randomly generated.

Updating the microVersionId field can be useful to:

- force updates in MongoDB when no other metadata changes

- detect a change when fields change but object version does not
  change e.g. when ingesting a putObjectTagging coming from S3C to
  Zenko

- manage concurrent updates, by adding a condition on the
  microVersionId and updating it for each metadata update to perform

In order for the change to be less intrusive, it is an optional field:
if ObjectMD.updateMicroVersionId() is not called by the client, the
metadata will not contain a microVersionId field. Clients will call
this function when needed on a case-by-case basis.