Use UTF-8 lexicographic order when filtering out of order keys
The default String comparison operators compare Strings using UTF-16 binary ordering.
Using this ordering codepoints in the BMP region after the UTF-16 surrogate pair region are incorrectly considered to be greater than codepoints outside the BMP.

Note that the default String <= is still used under the assumption that this type of keys will be rare and the <= will be faster than utf8Compare. This way the extra cost of utf8Compare only applies in the rare cases where these ordering issues do occur.

Signed-off-by: Pepijn Van Eeckhoudt <pepijn.vaneeckhoudt@datadobi.com>