bf: do not sort when searching
We change the default behavior of search by removing the sort directive.
Indeed some search queries may use other customer created index and
therefore cannot use the primary key _id, so it has to wait for
all results client side and sort them in memory which is highly
inefficient. Note that the sort might be useful when searching
so a long term fix shall have the sort optional.