bugfix: S3C-2804: Use ordered bulk write and refacto condition
BulkWrite was using a value of 1 to ordered which was
not correctly understood and lead to using unordered
batch write.
Refactor the catch of 11000 error to allow a more
specific catch and use a retry before sending back
an internal error in case there is a
race condition between multiple S3 connector on the
same object versionId.