ARSN-389 [squash] don't extend gaps with zero extra weight
Prevent extending an existing gap when the extension is zero
weight. This avoids creating useless zero-weight gaps when there is
concurrent invalidation of the gap being listed.

Note: It is still possible to create a gap smaller than the minimum
weight if the original gap has been invalidated, but it is less likely
than the previous odds of creating a zero-weight gap because the
original gap is usually already maxing out the skippable range.